feat: refactor all order-service & implement feedback apis

+++ backend/order-service/src/main/java/vn/edu/hust/order_service/dto/PricingResult.java
@@ -1,10 +1,12 @@
 package vn.edu.hust.order_service.dto;
 
+import java.math.BigDecimal;
+
 public record PricingResult(double weight,
-                            double shippingFee,
-                            double codFee,
-                            double insuranceFee,
-                            double surcharge,
-                            double totalPrice,
-                            double finalCollect) {
+                            BigDecimal shippingFee,
+                            BigDecimal codFee,
+                            BigDecimal insuranceFee,
+                            BigDecimal surcharge,
+                            BigDecimal totalPrice,
+                            BigDecimal finalCollect) {
 }
